+++ clinicadl/train/resume/resume_cli.py
@@ -1,22 +1,20 @@
 import click
 
-from clinicadl.utils import cli_param
+from clinicadl.config import arguments
+from clinicadl.config.options import (
+    cross_validation,
+    reproducibility,
+)
 
 
 @click.command(name="resume", no_args_is_help=True)
-@cli_param.argument.input_maps
-@cli_param.option_group.cross_validation.option(
-    "--split",
-    "-s",
-    type=int,
-    multiple=True,
-    help="Train the list of given splits. By default, all the splits are trained.",
-)
-def cli(input_maps_directory, split):
+@arguments.input_maps
+@cross_validation.split
+def cli(**kwargs):
     """Resume training job in specified maps.
 
     INPUT_MAPS_DIRECTORY is the path to the MAPS folder where training job has started.
     """
     from .resume import automatic_resume
 
-    automatic_resume(input_maps_directory, user_split_list=split)
+    automatic_resume(kwargs["input_maps_directory"], user_split_list=kwargs["split"])
